+++ src/core/transfer.rs
@@ -16,14 +16,14 @@ pub enum OfferType {
     Message(String),
     File {
         filename: String,
-        filesize: u32,
+        filesize: u64,
     },
     Directory {
         dirname: String,
         mode: String,
-        zipsize: u32,
-        numbytes: u32,
-        numfiles: u32,
+        zipsize: u64,
+        numbytes: u64,
+        numfiles: u64,
     },
 }
 
@@ -85,7 +85,7 @@ pub fn message(msg: &str) -> PeerMessage {
     PeerMessage::Offer(OfferType::Message(msg.to_string()))
 }
 
-pub fn offer_file(name: &str, size: u32) -> PeerMessage {
+pub fn offer_file(name: &str, size: u64) -> PeerMessage {
     PeerMessage::Offer(OfferType::File {
         filename: name.to_string(),
         filesize: size,
@@ -107,9 +107,9 @@ pub fn error_message(msg: &str) -> PeerMessage {
 pub fn offer_directory(
     name: &str,
     mode: &str,
-    compressed_size: u32,
-    numbytes: u32,
-    numfiles: u32,
+    compressed_size: u64,
+    numbytes: u64,
+    numfiles: u64,
 ) -> PeerMessage {
     PeerMessage::Offer(OfferType::Directory {
         dirname: name.to_string(),
